refactor: refactor test insights
