+++ packages/dashboard/src/insights/utils.ts
@@ -1,4 +1,11 @@
-import { Filters } from '../generated/graphql';
+import { GridRowsProp } from '@mui/x-data-grid';
+import { differenceInDays } from 'date-fns';
+import stringHash from 'string-hash';
+import {
+  FailedTestAggregate,
+  Filters,
+  FlakyTestAggregate,
+} from '../generated/graphql';
 
 export const getDateRange = (days: string) => {
   if (days === '0') return;
@@ -31,3 +38,33 @@ export const buildFilters = (
 
   return filters;
 };
+
+export function capitalizeFirstLetter(string: string) {
+  return string.charAt(0).toUpperCase() + string.slice(1);
+}
+
+export function convertTestsToRows(
+  tests: FailedTestAggregate[] | FlakyTestAggregate[]
+): GridRowsProp {
+  return tests.map((test) => {
+    return {
+      instanceId: stringHash(test.spec).toString(),
+      specName: test.spec,
+      firstFailedRun: test ? test : 'N/A',
+      lastFailedRun: test ? test : 'N/A',
+    };
+  });
+}
+
+export function getNumberOfDaysAgo(runDate: string) {
+  const daysAgo = differenceInDays(new Date(), new Date(runDate || 0)) + 1;
+  return `${daysAgo} days ago`;
+}
+
+export function getNumberOfTestRunsAgo(
+  numberOfTotalRuns: number,
+  runIndex: number
+) {
+  const testRunsAgo = numberOfTotalRuns - runIndex;
+  return `${testRunsAgo} test runs ago`;
+}
